raidemulator: Move query and clone helper methods to EmulatorCommon to deduplicate code
